scripts/requirements.txt
vosk>=0.3.42
requests>=2.28.0
tqdm>=4.65.0
faster-whisper>=0.10.0
nvidia-cublas-cu12
nvidia-cudnn-cu12
